Create tests for SLLI instruction
This commit creats an assembly file to test the correct
execution and decode of the SLLI instruction. To do so the Makefile was
changed to generate a new .hex file which contains an instruction that
doesn't follow the instruction spec. In this test case the processor
should throw a trap.